Moved plenty of code around to separate pure python implementation from future ones. This will change plenty of imports, which still needs to be fixed. Fortunately, this is a good foundation for getting all the tests fixed one again. Another step is to make the tests more flexible by allowing to run them with different object database easily.
