+++ gitdb/db/py/mem.py
